webtestpilot/pyproject.toml
[project]
name = "webtestpilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baml-py>=0.201.0",
    "matplotlib>=3.10.3",
    "openai>=1.95.1",
    "pillow>=11.3.0",
    "playwright>=1.53.0",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
    "pyyaml>=6.0.2",
    "ruff>=0.12.3",
    "zss>=1.2.0",
]
